+++ CHANGELOG.md
@@ -8,9 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2022-10-27
+
+### Added
+
+F004 - Islanding script
+
 ## [1.0.0] - 2022-10-24
 
 ### Added
 
 F001 - Workspace setup: sandbox scaffolding, boilerplate code, physics debug.
 F002 - Apple II sprites and tilemap
+F003 - Use URL param & prod build basics